requirements.txt
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.9
MarkupSafe==1.1.1
requests==2.24.0
urllib3==1.25.9
Werkzeug==1.0.1
flask-redis==0.4.0
PyJWT==1.7.1

# Tests
pytest-flask==1.0.0
mock==4.0.2
ipdb
